refactor: Change origins api.olivedb.info and sort routes
