+++ backend/main.py
@@ -1,12 +1,13 @@
 from fastapi import FastAPI
-from api.products import router
+from api.products import products_router
+from api.price_by_date import price_by_date_router
 from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI()
 
 origins = [
-    "http://oliveDB.info",
-    "https://oliveDB.info",
+    "http://api.olivedb.info",
+    "https://api.olivedb.info",
 ]
 
 app.add_middleware(
@@ -17,4 +18,11 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-app.include_router(router, prefix="/api")
+
+@app.get("/")
+def get_names():
+    return {"Welcom": "api.olive.db"}
+
+
+app.include_router(products_router, prefix="/products")
+app.include_router(price_by_date_router, prefix="/price-by-date")
